Update and remove some crusty scripts
svn path=/trunk/; revision=784

+++ scripts/upload-doc.sh
@@ -1,2 +1,9 @@
 #!/bin/bash
+
+APP="conduit/start_conduit.py"
+if [ ! -f $APP ] ; then
+    echo "ERROR: Must be run from top directory"
+    exit 1
+fi
+
 rsync -Ptz doc/*.{html,gif,png,py,js,css} root@greenbirdsystems.com:/var/www/conduit-project.org/doc/conduit
